+++ inc/classes/layer.js
@@ -25,14 +25,12 @@
     let Port = require('./port');
     let Label = require('./label');
     let Volume = require('./volume');
-    let RunAfter = require('./runAfter');
     let VolumeFrom = require('./volumeFrom');
     let Environment = require('./environment');
 
     let fs = require('fs');
     let _ = require('lodash');
     let path = require('path');
-    let async = require('async');
     let bytes = require('bytes');
     let touch = require('touch');
     let mkdirp = require('mkdirp');
@@ -91,14 +89,6 @@
             if (originalObject.volumes) {
                 _.forEach(originalObject.volumes, function (volume) {
                     this[objectSymbol].volumes.push(new Volume(this, volume));
-                }, this);
-            }
-
-            // Turn the runAfters from in the object into RunAfter objects
-            this[objectSymbol].runAfter = [];
-            if (originalObject.runAfter) {
-                _.forEach(originalObject.runAfter, function (runAfter) {
-                    this[objectSymbol].runAfter.push(new RunAfter(this, runAfter));
                 }, this);
             }
 
@@ -178,7 +168,7 @@
          * @returns {Boolean}
          */
         get dataOnly() {
-            return this[objectSymbol].dataOnly || false;
+            return false;
         }
 
         /**
@@ -406,21 +396,12 @@
         }
 
         /**
-         * Gets the things to run after this layer is finished (runOnly layers only).
-         *
-         * @returns {RunAfter[]}
-         */
-        get runAfter() {
-            return this[objectSymbol].runAfter;
-        }
-
-        /**
          * Returns if this layer is a run only layer which means it's intended to be run by itself only.
          *
          * @returns {Boolean}
          */
         get runOnly() {
-            return this[objectSymbol].runOnly || false;
+            return false;
         }
 
         /**
@@ -481,27 +462,6 @@
             }
 
             return value;
-        }
-
-        /**
-         * See's if this layer can be run or not.
-         *
-         * @returns {Promise}
-         */
-        canRun() {
-            return new Promise(function (resolve, reject) {
-                if (!this.runOnly) {
-                    return reject(new Error('Cannot run this layer because it\'s not set as runOnly!'));
-                }
-
-                this.application.areLayersUp(this.dependentLayers).then(function (areUp) {
-                    if (!areUp) {
-                        return resolve(false);
-                    }
-
-                    resolve(true);
-                }).catch(reject);
-            }.bind(this));
         }
 
         /**
@@ -647,102 +607,6 @@
                         }).catch(reject);
                     });
                 }).catch(reject);
-            }.bind(this));
-        }
-
-        /**
-         * This runs a layer for the application that's set as runOnly. The stdin is attached and allows interaction with the container.
-         *
-         * @param {Object|Array} options - options passed in from the user
-         * @returns {Promise}
-         */
-        run(options) {
-            let self = this;
-
-            return new Promise(function (resolve, reject) {
-                this.container.remove(function () {
-                    let dOpts = self.dockerOptions;
-
-                    // Change our specific options for Docker
-                    dOpts.Cmd = self.command.concat(options instanceof Array ? options : options._raw.slice(options._raw.indexOf(self.name) + 1));
-                    dOpts.AttachStdin = true;
-                    dOpts.AttachStdout = true;
-                    dOpts.AttachStderr = true;
-                    dOpts.Tty = true;
-                    dOpts.OpenStdin = true;
-                    dOpts.StdinOnce = false;
-
-                    brain.docker.getImage(self.image).inspect(function (err) {
-                        if (err) {
-                            pullAndUp();
-                        } else {
-                            checkLayers();
-                        }
-                    });
-
-                    let layersNeeded = [];
-
-                    function pullAndUp() {
-                        self.pull().then(checkLayers).catch(reject);
-                    }
-
-                    function checkLayers() {
-                        self.canRun().then(function (canRun) {
-                            if (canRun) {
-                                return runIt();
-                            }
-
-                            async.eachSeries(self.dependentLayers, function (layerName, next) {
-                                let layer = self.application.getLayer(layerName);
-
-                                if (!layer) {
-                                    return next(new Error(`There is no layer called ${layerName} for this application!`));
-                                }
-
-                                layer.isUp().then(function (isUp) {
-                                    if (!isUp) {
-                                        layersNeeded.push(layer);
-                                    }
-
-                                    next();
-                                }).catch(reject);
-                            }, function (err) {
-                                if (err) {
-                                    return reject(err);
-                                }
-
-                                async.eachSeries(layersNeeded, function (layer, next) {
-                                    layer.up(options).then(() => next()).catch(next);
-                                }, (err) => err ? reject(err) : runIt());
-                            });
-                        }).catch(reject);
-                    }
-
-                    function runIt() {
-                        brain.run(dOpts).then(function (err) {
-                            if (err) {
-                                return reject(err);
-                            }
-
-                            async.eachSeries(layersNeeded, function (layer, next) {
-                                options.rm = true;
-                                layer.down(options).then(() => next()).catch(next);
-                            }, function (err) {
-                                if (err) {
-                                    return reject(err);
-                                }
-
-                                if (self.runAfter) {
-                                    async.eachSeries(self.runAfter, function (ra, next) {
-                                        ra.layer.run(ra.command).then(() => next()).catch(next);
-                                    }, (err) => err ? reject(err) : resolve());
-                                } else {
-                                    resolve();
-                                }
-                            });
-                        }).catch(reject);
-                    }
-                });
             }.bind(this));
         }